major refactor to fastapi, setup realtime web search agent
