+++ build_assets/macOS/preinstall_ollama.sh
@@ -7,5 +7,5 @@ chmod +x ollama
 sudo mv ollama /usr/local/bin/
 
 nohup /usr/local/bin/ollama serve > /dev/null 2>&1 &
-/usr/local/bin/ollama pull llama3.1
+/usr/local/bin/ollama pull llama3.2:3b
 /usr/local/bin/ollama pull nomic-embed-text
