+++ build_assets/macOS/welcome.html
@@ -1,44 +1,57 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Welcome to MORagents v0.1.0 Installer</title>
-  <style>
-    body {
-      font-family: Arial, sans-serif;
-      margin: 20px;
-    }
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Welcome to MORagents v0.1.1 Installer</title>
+    <style>
+      body {
+        font-family: Arial, sans-serif;
+        margin: 20px;
+      }
 
-    h1 {
-      color: #333;
-    }
+      h1 {
+        color: #333;
+      }
 
-    p {
-      color: #666;
-      line-height: 1.5;
-    }
+      p {
+        color: #666;
+        line-height: 1.5;
+      }
 
-    ul {
-      margin-top: 20px;
-    }
+      ul {
+        margin-top: 20px;
+      }
 
-    li {
-      margin-bottom: 10px;
-    }
-  </style>
-</head>
-<body>
-  <h1>Welcome to MORagents v0.1.0 Installer</h1>
-  <p>Thank you for choosing to install MORagents on your system. This installer will guide you through the process of setting up MORagents and its dependencies.</p>
-  <p>The installer will perform the following steps:</p>
-  <ul>
-    <li>Check the installed Python version and install Python 3.12.0 if necessary.</li>
-    <li>Check if Docker is installed and install Docker Desktop if needed.</li>
-    <li>Check if Ollama is installed and install Ollama if needed.</li>
-    <li>Install the MORagents application.</li>
-  </ul>
-  <p>Please note that during the installation process, you may be prompted to enter your system password to authorize the installation of required components.</p>
-  <p>Click "Continue" to proceed with the installation.</p>
-</body>
+      li {
+        margin-bottom: 10px;
+      }
+    </style>
+  </head>
+  <body>
+    <h1>Welcome to MORagents v0.1.1 Installer</h1>
+    <p>
+      Thank you for choosing to install MORagents on your system. This installer
+      will guide you through the process of setting up MORagents and its
+      dependencies.
+    </p>
+    <p>The installer will perform the following steps:</p>
+    <ul>
+      <li>
+        Check the installed Python version and install Python 3.12.0 if
+        necessary.
+      </li>
+      <li>
+        Check if Docker is installed and install Docker Desktop if needed.
+      </li>
+      <li>Check if Ollama is installed and install Ollama if needed.</li>
+      <li>Install the MORagents application.</li>
+    </ul>
+    <p>
+      Please note that during the installation process, you may be prompted to
+      enter your system password to authorize the installation of required
+      components.
+    </p>
+    <p>Click "Continue" to proceed with the installation.</p>
+  </body>
 </html>
