+++ submodules/moragents_dockers/agents/Dockerfile-apple
@@ -12,18 +12,11 @@ RUN apt-get update && apt-get install -y gcc g++ procps && rm -rf /var/lib/apt/l
 # Install Python dependencies
 RUN python3 -m pip install --no-cache-dir --upgrade pip && \
     python3 -m pip install --no-cache-dir --upgrade -r requirements.txt
-    
-COPY download_model.py .
 
-COPY model_config.py .
-
-RUN python3 download_model.py
-
-copy . .
+COPY . .
 
 # Expose the port your application listens on
 EXPOSE 5000
-
 
 # Set the environment variable for Flask
 ENV FLASK_APP=src/app.py
