--- submodules/moragents_dockers/agents/model_config.py
@@ -1,13 +1,0 @@
-import logging
-
-# Logging configuration
-logging.basicConfig(level=logging.INFO)
-
-
-# Configuration object
-class Config:
-    # Model configuration
-    MODEL_NAME = "meetkai/functionary-small-v2.4-GGUF"
-    MODEL_REVISION = "functionary-small-v2.4.Q4_0.gguf"
-    MODEL_PATH = "model/" + MODEL_REVISION
-    DOWNLOAD_DIR = "model"
