+++ submodules/moragents_dockers/agents/src/agents/crypto_data/routes.py
@@ -1,0 +1,14 @@
+import logging
+
+from flask import Blueprint, request, jsonify
+
+crypto_data_agent_bp = Blueprint('crypto_data_agent', __name__)
+logger = logging.getLogger(__name__)
+
+@crypto_data_agent_bp.route('/process_data', methods=['POST'])
+def process_data():
+    logger.info("Data Agent: Received process_data request")
+    data = request.get_json()
+    # Implement your data processing logic here
+    response = {"status": "success", "message": "Data processed"}
+    return jsonify(response)
