+++ submodules/moragents_dockers/agents/src/agents/rag/config.py
@@ -1,0 +1,10 @@
+import logging
+
+# Logging configuration
+logging.basicConfig(level=logging.INFO)
+
+
+# Configuration object
+class Config:
+    MAX_FILE_SIZE = 5 * 1024 * 1024  # 5 MB
+    MAX_LENGTH = 16 * 1024 * 1024
