+++ submodules/moragents_dockers/agents/src/models/messages.py
@@ -1,0 +1,11 @@
+from pydantic import BaseModel
+
+
+class ChatMessage(BaseModel):
+    role: str
+    content: str
+
+
+class ChatRequest(BaseModel):
+    user_id: str
+    prompt: ChatMessage
