+++ submodules/moragents_dockers/docker-compose.yml
@@ -2,7 +2,7 @@ version: "3.8"
 
 services:
   agents:
-    image: lachsbagel/moragents_dockers-agents:amd64-0.1.0
+    image: lachsbagel/moragents_dockers-agents:amd64-0.1.1
     build:
       dockerfile: Dockerfile
       context: ./agents
@@ -18,7 +18,7 @@ services:
       - BASE_URL=http://host.docker.internal:11434
 
   nginx:
-    image: lachsbagel/moragents_dockers-nginx:amd64-0.1.0
+    image: lachsbagel/moragents_dockers-nginx:amd64-0.1.1
     build:
       context: ./frontend
       dockerfile: Dockerfile
