+++ submodules/moragents_dockers/frontend/package.json
@@ -1,7 +1,7 @@
 {
   "name": "Morpheus AI",
   "private": true,
-  "version": "0.1.0",
+  "version": "0.1.1",
   "scripts": {
     "dev": "next dev",
     "build": "next build",
